requirements.txt
# Core ML and Data Analysis
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
tensorflow==2.15.0
python-dateutil==2.8.2
requests==2.31.0

# Web Framework
Flask==3.0.0
Flask-SocketIO==5.3.6
Werkzeug==3.0.1

# News Sentiment Analysis
transformers==4.36.2
torch==2.1.1
vaderSentiment==3.3.2
beautifulsoup4==4.12.2
newsapi-python==0.2.7
textblob==0.17.1

# Whale Tracking
web3==6.11.3
python-dotenv==1.0.0

# Crypto Trading and Data
ccxt==4.1.77

# MSSQL Database Integration (Mac Compatible)
pyodbc==4.0.39

# Model Cache System
# (pickle, hashlib, json, os, datetime are built-in Python modules)
Flask-Login==0.6.3

# Deep Q-Learning and Reinforcement Learning
gym==0.26.2
stable-baselines3==2.0.0

# Enhanced Data Science and Visualization
plotly==5.15.0
dash==2.11.1

# Additional TensorFlow Tools
tensorboard==2.13.0

# Advanced Trading Algorithms
quantlib==1.31
